BUG/MINOR: connection: fix loop iter on connection takeover
The loop always missed one iteration due to the incrementation done on
the for check. Move the incrementation on the loop last statement to fix
this behaviour.

This bug has a very limited impact, not at all visible to the user, but
could be backported to 2.2.

(cherry picked from commit 5f1ded5629d046124827087c41e765d5362c87cf)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>